refactor: split example package code into multiple modules
